split decoders to separate files
