+++ src/decoders/index.ts
@@ -1,0 +1,3 @@
+import { decodeSystemInstruction } from "./system";
+
+export { decodeSystemInstruction };
